+++ src/ios/OSFirebaseCloudMessaging.swift
@@ -88,8 +88,8 @@ class OSFirebaseCloudMessaging: CDVPlugin {
         self.callbackId = command.callbackId
         guard
             let badge = command.arguments[0] as? Int,
-            let body = command.arguments[1] as? String,
-            let title = command.arguments[2] as? String
+            let title = command.arguments[1] as? String,
+            let body = command.arguments[2] as? String
         else {
             self.sendResult(result: "", error:FirebaseMessagingErrors.sendNotificationsError as NSError, callBackID: self.callbackId)
             return
@@ -208,9 +208,13 @@ extension OSFirebaseCloudMessaging: FirebaseMessagingCallbackProtocol {
 // MARK: - OSFirebaseMessagingLib's FirebaseMessagingEventProtocol Methods
 extension OSFirebaseCloudMessaging: FirebaseMessagingEventProtocol {
     func event(_ event: FirebaseEventType, data: String) {
-        var eventName = "notificationClick"
-        if case let .trigger(notification) = event {
-            eventName = notification == .silentNotification ? "silentNotification" : "defaultNotification"
+        var eventName = ""
+        
+        switch event {
+        case .click(type: let type):
+            eventName = type == .notification ? "notificationClick" : "internalRouteActionClick"
+        case .trigger(notification: let notification):
+            eventName = notification == .silentNotification ? "silentNotification" : "defaultNotification"        
         }
         
         self.trigger(event: eventName, data: data)