+++ src/ios/Utils/OSFCMTokenType.swift
@@ -1,5 +1,0 @@
-/// Indicates the type of token to be requested
-public enum OSFCMTokenType {
-    case fcm
-    case apns
-}